Create a virtual environment; add .gitignore and requirements.txt
A .gitignore in a git repo will tell git which files and directories NOT
to track.

__pycache__:
A directory that holds artifacts from running a Python script. They're
safe to delete and hold no value to the codebase.

venv:
This is a special folder that holds a "virtual environment" on your
local machine. It can be named anything, but "venv" or ".venv" is the
accepted convention. A virtual environment is something a developer
chooses to create for a project, allowing them to install packages (like
numpy or pandas) for ONLY that project.

    To create a virtual environment, run:
    ```
    python -m venv venv
    ```

    `python` invokes the Python interpreter.
    `-m` tells Python the next argument is a module
    `venv` is the virtual environment module
    `venv` is the name you're giving your virtual environment

    Then you can "activate" your virtual environment like this:
    ```
    source venv/bin/activate
    ```

    Once it's activated, install Python dependencies like this:
    ```
    pip install numpy pandas

    # or

    pip install -r requirements.txt
    ```

    You can get out of your venv by running:
    ```
    deactivate
    ```

.DS_Store:
A completely useless file generated by macOS. Safe to delete
(recommended).